доработка замечаний
